revocation: Move `CertRevocationList` to its own module.
